fix(wiki): enhance documentation
