+++ wiki/docs/Endpoints/Metrics.md
@@ -4,6 +4,13 @@ The application exposes Prometheus metrics at the `/metrics` endpoint. This endp
 
 By default, this endpoint is available on Port `9120`, but can be configured using the `METRICS_PORT` environment variable, see [App Settings](../App-Settings.md#available-settings).
 
+## Available Metrics
+
+See the following Source Code to find out about the currently available metrics:
+```go title="Prometheus Collectors"
+--8<-- "internal/prometheus/collectors.go"
+```
+
 ## Grafana Dashboard
 
 An example for a Grafana Dashboard can be found at [Grafana Dashboard #583](https://github.com/kimdre/doco-cd/discussions/583).
