+++ wiki/docs/external-secrets/1password.md
@@ -1,5 +1,9 @@
 # 1Password
 
+!!! info
+    The start time and memory usage of the doco-cd container, as well as the runtime of a job, can increase significantly when using this secret provider.
+
+## Environment Variables
 
 To use 1Password, you need to set the following environment variables:
 
@@ -9,11 +13,7 @@ To use 1Password, you need to set the following environment variables:
 | `SECRET_PROVIDER_ACCESS_TOKEN`      | Access token of a service account, see [the docs](https://developer.1password.com/docs/service-accounts/security/) and [here](https://developer.1password.com/docs/sdks/setup-tutorial/#part-1-set-up-a-1password-service-account) |
 | `SECRET_PROVIDER_ACCESS_TOKEN_FILE` | Path to the file containing the access token inside the container                                                                                                                                                                  |
 
-!!! note
-    The start time and memory usage of the doco-cd container, as well as the runtime of a job, can increase significantly when using this secret provider.
-
-
-**Deployment configuration**
+## Deployment configuration
 
 Add a mapping/reference between the environment variable you want to set in the docker compose project/stack and the URI to the secret in 1Password.
 
@@ -28,10 +28,11 @@ To get a one-time password, append the `?attribute=otp` query parameter to a sec
 !!! warning
     Machine accounts can only access vaults for which you have granted read permissions during creation. The default `Personal` vault can't be access by machine accounts!
 
+### Example
 
 For example in your `.doco-cd.yml`:
 
-```yaml
+```yaml title=".doco-cd.yml"
 name: myapp
 external_secrets:
   DB_PASSWORD: "op://vault/item/field"
