+++ wiki/docs/external-secrets/openbao.md
@@ -1,5 +1,6 @@
 # OpenBao
 
+## Environment Variables
 
 To use OpenBao, you need to set the following environment variables:
 
@@ -10,25 +11,34 @@ To use OpenBao, you need to set the following environment variables:
 | `SECRET_PROVIDER_ACCESS_TOKEN`      | Access token for authenticating with the secret provider        |
 | `SECRET_PROVIDER_ACCESS_TOKEN_FILE` | Path to a file containing the access token inside the container |
 
-**Deployment configuration**
+## Deployment configuration
 
 Add a mapping/reference between the environment variable you want to set in the docker compose project/stack and the reference to the key-value secret in OpenBao.
 
 By default, the root namespace is used (`root` or `/`), but you can specify a different namespace by adding it as the first part of the reference.
 
-- A valid key-value secret reference should use the syntax: `kv:<namespace(optional)>:<secretEngine>:<secretName>:<key>`
-- A valid PKI certificate reference should use the syntax: `pki:<namespace(optional)>:<secretEngine>:<commonName>`
+- A valid key-value secret reference should use the syntax: 
+  ```
+  kv:<namespace(optional)>:<secretEngine>:<secretName>:<key>
+  ```
+- A valid PKI certificate reference should use the syntax:
+  ```
+  pki:<namespace(optional)>:<secretEngine>:<commonName>
+  ```
 
 Examples of valid references:
+
 - `kv:prod-secrets:db-prod:username` &rarr; Fetches the `username` key from the `db-prod` key-value secret in the `prod-secrets` secret engine in the `root` namespace.
 - `kv:root:prod-secrets:db-prod:username` &rarr; Same as above, explicitly specifying the `root` namespace.
 - `kv:my-namespace:secret:api-keys:stripe` &rarr; Fetches the `stripe` key from the `api-keys` secret in the `secret` key-value secret engine in the `my-namespace` namespace.
 - `pki:certs:myapp.example.com` &rarr; Fetches the certificate for the common name `myapp.example.com` from the `certs` pki secret engine in the `root` namespace.
 - `pki:my-namespace:certs:myapp.example.com` &rarr; Fetches the certificate for the common name `myapp.example.com` from the `certs` pki secret engine in the `my-namespace` namespace.
 
+### Example
+
 For example in your `.doco-cd.yml`:
 
-```yaml
+```yaml title=".doco-cd.yml"
 name: myapp
 external_secrets:
   DB_USERNAME: kv:secret:db-prod:username
@@ -38,8 +48,7 @@ external_secrets:
 
 To use the certificate in your compose file, you can pass the value to a compose config:
 
-```yaml
-# docker-compose.yml
+```yaml title="docker-compose.yml"
 configs:
   myapp-example-com.crt:
     #environment: CERT  # Either pass the variable via the environment like this (without a $ sign)
